pyproject.toml
[tool.poetry]
name = "cluster-server"
version = "0.1.0"
description = ""
authors = ["ragnarok56 <kevin.nacios@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
ansible = "2.10"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
